refactor: split main.go and appconstants into domain-specific files
Split main.go (~1280 lines) into cmd_gen.go, cmd_validate.go, cmd_config.go,
cmd_deps.go, and cmd_cache.go, leaving main.go as a lean entry point.

Split appconstants/constants.go (~840 lines) into domain-specific files:
config.go, errors.go, files.go, flags.go, git.go, testing.go, themes.go,
validation.go. Each file contains only constants for its domain.
