+++ appconstants/flags.go
@@ -1,0 +1,26 @@
+// Package appconstants provides common constants used throughout the application.
+package appconstants
+
+// CLI flag and command names.
+const (
+	// FlagFormat is the format flag name.
+	FlagFormat = "format"
+	// FlagOutputDir is the output-dir flag name.
+	FlagOutputDir = "output-dir"
+	// FlagOutputFormat is the output-format flag name.
+	FlagOutputFormat = "output-format"
+	// FlagOutput is the output flag name.
+	FlagOutput = "output"
+	// FlagRecursive is the recursive flag name.
+	FlagRecursive = "recursive"
+	// FlagIgnoreDirs is the ignore-dirs flag name.
+	FlagIgnoreDirs = "ignore-dirs"
+	// FlagCI is the CI mode flag name.
+	FlagCI = "ci"
+
+	// CommandPin is the pin command name.
+	CommandPin = "pin"
+
+	// CacheStatsKeyDir is the cache stats key for directory.
+	CacheStatsKeyDir = "cache_dir"
+)
