+++ cmd_validate.go
@@ -1,0 +1,63 @@
+// Package main provides the validate and schema commands.
+package main
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+
+	"github.com/ivuorinen/gh-action-readme/appconstants"
+	"github.com/ivuorinen/gh-action-readme/internal"
+	"github.com/ivuorinen/gh-action-readme/internal/helpers"
+)
+
+func newValidateCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "validate",
+		Short: "Validate action.yml files and optionally autofill missing fields.",
+		Run:   wrapHandlerWithErrorHandling(validateHandler),
+	}
+}
+
+func newSchemaCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "schema",
+		Short: "Show the action.yml schema info.",
+		Run:   schemaHandler,
+	}
+}
+
+func validateHandler(_ *cobra.Command, _ []string) error {
+	currentDir, err := helpers.GetCurrentDir()
+	if err != nil {
+		return fmt.Errorf("unable to determine current directory: %w", err)
+	}
+
+	generator := internal.NewGenerator(globalConfig)
+	actionFiles, err := generator.DiscoverActionFilesWithValidation(
+		currentDir,
+		true,
+		globalConfig.IgnoredDirectories,
+		"validation",
+	) // Recursive for validation
+	if err != nil {
+		return fmt.Errorf(appconstants.ErrFailedToDiscoverActionFiles, err)
+	}
+
+	// Validate the discovered files
+	if err := generator.ValidateFiles(actionFiles); err != nil {
+		return fmt.Errorf("validation failed for %d files: %w", len(actionFiles), err)
+	}
+
+	generator.Output.Success("\nAll validations passed successfully!")
+
+	return nil
+}
+
+func schemaHandler(_ *cobra.Command, _ []string) {
+	output := internal.NewColoredOutput(globalConfig.Quiet)
+	if globalConfig.Verbose {
+		output.Info("Using schema: %s", globalConfig.Schema)
+	}
+	output.Printf("Schema: schemas/action.schema.json (replaceable, editable)")
+}
